Major update rewritten OOP style, customization panel, category buttons many fixes  and improvements.
Customization panel shown with right click on the desired marker, choice to use default setting and customized settings
Category buttons are automatically updated and shown on the side of the selector. Marker objects are sorted alphabetically and need to start with "somecategory_" to register as part of it. Category button make the selector scroll to the first corresponding. Nice highlight of cat button to show were we're at.

Change name DrawFromImage.lua to gui_draw_marker_from_image.lua